Normalize KeyframeTrack, apply MrDoob eslint style, merge KeyframeTrackPrototype and KeyframeTrackConstructor in a single file. Update import statement of tracks
