+++ src/animation/tracks/BooleanKeyframeTrack.js
@@ -1,6 +1,5 @@
 import { InterpolateDiscrete } from '../../constants';
-import { KeyframeTrackPrototype } from '../KeyframeTrackPrototype';
-import { KeyframeTrackConstructor } from '../KeyframeTrackConstructor';
+import { KeyframeTrack } from '../KeyframeTrack';
 
 /**
  *
@@ -14,12 +13,11 @@ import { KeyframeTrackConstructor } from '../KeyframeTrackConstructor';
 
 function BooleanKeyframeTrack( name, times, values ) {
 
-	KeyframeTrackConstructor.call( this, name, times, values );
+	KeyframeTrack.call( this, name, times, values );
 
 }
 
-BooleanKeyframeTrack.prototype =
-		Object.assign( Object.create( KeyframeTrackPrototype ), {
+BooleanKeyframeTrack.prototype = Object.assign( Object.create( KeyframeTrack.prototype ), {
 
 	constructor: BooleanKeyframeTrack,
 
@@ -36,6 +34,5 @@ BooleanKeyframeTrack.prototype =
 	// computes "firstValue ^ isOdd( index )".
 
 } );
-
 
 export { BooleanKeyframeTrack };
