+++ src/animation/tracks/QuaternionKeyframeTrack.js
@@ -1,7 +1,6 @@
 import { InterpolateLinear } from '../../constants';
-import { KeyframeTrackPrototype } from '../KeyframeTrackPrototype';
+import { KeyframeTrack } from '../KeyframeTrack';
 import { QuaternionLinearInterpolant } from '../../math/interpolants/QuaternionLinearInterpolant';
-import { KeyframeTrackConstructor } from '../KeyframeTrackConstructor';
 
 /**
  *
@@ -14,12 +13,11 @@ import { KeyframeTrackConstructor } from '../KeyframeTrackConstructor';
 
 function QuaternionKeyframeTrack( name, times, values, interpolation ) {
 
-	KeyframeTrackConstructor.call( this, name, times, values, interpolation );
+	KeyframeTrack.call( this, name, times, values, interpolation );
 
 }
 
-QuaternionKeyframeTrack.prototype =
-		Object.assign( Object.create( KeyframeTrackPrototype ), {
+QuaternionKeyframeTrack.prototype = Object.assign( Object.create( KeyframeTrack.prototype ), {
 
 	constructor: QuaternionKeyframeTrack,
 
@@ -29,16 +27,14 @@ QuaternionKeyframeTrack.prototype =
 
 	DefaultInterpolation: InterpolateLinear,
 
-	InterpolantFactoryMethodLinear: function( result ) {
+	InterpolantFactoryMethodLinear: function ( result ) {
 
-		return new QuaternionLinearInterpolant(
-				this.times, this.values, this.getValueSize(), result );
+		return new QuaternionLinearInterpolant( this.times, this.values, this.getValueSize(), result );
 
 	},
 
 	InterpolantFactoryMethodSmooth: undefined // not yet implemented
 
 } );
-
 
 export { QuaternionKeyframeTrack };
