+++ src/animation/tracks/VectorKeyframeTrack.js
@@ -1,5 +1,4 @@
-import { KeyframeTrackPrototype } from '../KeyframeTrackPrototype';
-import { KeyframeTrackConstructor } from '../KeyframeTrackConstructor';
+import { KeyframeTrack } from '../KeyframeTrack';
 
 /**
  *
@@ -13,12 +12,11 @@ import { KeyframeTrackConstructor } from '../KeyframeTrackConstructor';
 
 function VectorKeyframeTrack( name, times, values, interpolation ) {
 
-	KeyframeTrackConstructor.call( this, name, times, values, interpolation );
+	KeyframeTrack.call( this, name, times, values, interpolation );
 
 }
 
-VectorKeyframeTrack.prototype =
-		Object.assign( Object.create( KeyframeTrackPrototype ), {
+VectorKeyframeTrack.prototype = Object.assign( Object.create( KeyframeTrack.prototype ), {
 
 	constructor: VectorKeyframeTrack,
 
@@ -29,6 +27,5 @@ VectorKeyframeTrack.prototype =
 	// DefaultInterpolation is inherited
 
 } );
-
 
 export { VectorKeyframeTrack };
